more secure symbol encryption
Before, the encryption I used was basically just XORing a mask over the symbols. This lead to certain issues. For instance, if two symbols both started with the same letter, the first two characters of their encrypted names would be the same. This could allow for language analysis and eventually to decryption of data.