output easier-to-parse string obfuscation
Using large byte arrays like []byte{1, 2, ...} makes source files a LOT harder to parse than using "\xab\x32\x5a...", probably because there are so many more tokens. When running the obfuscator on a large piece of binary data stored in a string, the obfuscator used to spend loads of time and allocate ~20GB of memory. Now it uses closer to 2GB of memory and takes way less time for the same task.